Restore "default:" section, so setting of gcc_dir works.
SPECCPU config files have three section (types), the header (or default) section,
named sections, and the HASH section. The header section is the first, and most
attempts to modify runcpu occur in it. Those options can also be entered in
an explicitly labeled "default:" section. In cpu2017_tx2_gcc10, several named
sections occur after the implicit header section. Therefore, the attempt to define
a default for gcc_dir in the preceeding "fpspeed:" named section is not correct.

For more details see:
  Config Files - CPU 2017 https://www.spec.org/cpu2017/Docs/config.html